refactor: 봉사활동 공지사항 이름 변경 약간의 코드 수정
