--- pages/volunteer/detail/announcements/announcement/index.js
@@ -1,10 +1,0 @@
-import Announcement from "../../../../../src/components/units/volunteer/announcements/announcement/Announcement.container";
-
-
-export default function AnnouncementPage() {
-  return(
-    <>
-      <Announcement />
-    </>
-  )
-}
