--- pages/volunteer/detail/announcements/index.js
@@ -1,9 +1,0 @@
-import Announcements from "../../../../src/components/units/volunteer/announcements/Announcements.container";
-
-export default function AnnouncementsPage() {
-  return(
-    <>
-      <Announcements />
-    </>
-  )
-};
