+++ pages/volunteer/detail/notices/index.js
@@ -1,0 +1,9 @@
+import Notices from "../../../../src/components/units/volunteer/notices/notices.container";
+
+export default function AnnouncementsPage() {
+  return(
+    <>
+      <Notices />
+    </>
+  )
+};
