+++ pages/volunteer/detail/notices/notice/index.js
@@ -1,0 +1,10 @@
+import Notice from "../../../../../src/components/units/volunteer/notices/notice/notice.container";
+
+
+export default function AnnouncementPage() {
+  return(
+    <>
+      <Notice />
+    </>
+  )
+}
