+++ src/components/units/volunteer/notices/notice/component/LikeImage.js
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import Image from 'next/image';
-import * as S from '../Announcement.styles'; // LikeImage 컴포넌트의 스타일 파일 경로에 따라 수정해야 할 수도 있습니다.
+import * as S from '../notice.styles'; // LikeImage 컴포넌트의 스타일 파일 경로에 따라 수정해야 할 수도 있습니다.
 
 const LikeImage = ({ initialSrc, alt }) => {
   const [imgSrc, setImgSrc] = useState(initialSrc);
